Master/SelfDeleteFix
Updated RemoveUserModal to not allow a user to delete his own account, as that had some unwanted side effects.

+++ JobAgentSolutions/BlazorWebsite/Shared/Components/Modals/UserAccessModals/RemoveUserModal.razor.cs
@@ -2,14 +2,17 @@
 using JobAgentClassLibrary.Common.Users;
 using JobAgentClassLibrary.Common.Users.Entities;
 using Microsoft.AspNetCore.Components;
+using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.JSInterop;
 using System;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace BlazorWebsite.Shared.Components.Modals.UserAccessModals
 {
     public partial class RemoveUserModal : ComponentBase
     {
+        [CascadingParameter] private Task<AuthenticationState> AuthState { get; set; }
         [Parameter] public int Id { get; set; }
         [Inject] protected IRefreshProvider RefreshProvider { get; set; }
         [Inject] protected IJSRuntime JSRuntime { get; set; }
@@ -17,6 +20,8 @@ namespace BlazorWebsite.Shared.Components.Modals.UserAccessModals
 
         private string _errorMessage = "";
         private bool _isProcessing = false;
+        private string _sessionUserEmail;
+
 
         private async Task OnClick_RemoveJobPage(int id)
         {
@@ -24,10 +29,24 @@ namespace BlazorWebsite.Shared.Components.Modals.UserAccessModals
             {
                 _isProcessing = true;
 
-                IUser user = new User()
+                var session = await AuthState;
+
+                foreach (var sessionClaim in session.User.Claims)
                 {
-                    Id = id
-                };
+                    if (sessionClaim.Type == ClaimTypes.Email)
+                    {
+                        _sessionUserEmail = sessionClaim.Value;
+                    }
+                }
+
+
+                var user = await UserService.GetUserByIdAsync(id);
+
+                if(user.Email == _sessionUserEmail)
+                {
+                    _errorMessage = "Du kan ikke slette din egen bruger.";
+                    return;
+                }
 
                 var result = await UserService.RemoveAsync(user);
